HealthQualityImprovement_StoryTask1&2
Specific Aim: Using an Excel spreadsheet to develop a graph and a table to visualize and interpret the key findings using Tableau. The detailed assessment requirements were described as the readme.
Task1: It was generated a bar graph of breast cancer screening data in 2017, and the graph is grouped by all eight regions.
Task2: It was generated a highlighted data table to represent all six screening measures in both 2017 and 2018, and the data is grouped by all eight regions.
Besides two sheets, it was also generated a Story to integrate two sheets with interpretation and explanation regarding the key findings from two sheets in the Tableau file.